build(cli): wire the pipeline together and put diagnostics on stderr
Completes the port: resolve repo and host, authenticate, fetch, filter,
render, write.

Diagnostics now go to stderr unconditionally. The original printed its
progress line to stdout and then suppressed it whenever the output target
was stdout, which made `-o - | glow` work by luck rather than by
construction -- add a second status message and the pipe starts producing
corrupt Markdown. It also emitted ANSI escapes regardless of destination,
so a redirected log was full of colour codes; colour is now applied only
when stderr is a terminal.

Files are written through a temporary in the same directory and renamed.
People re-run this over a digest they are currently reading, and a partial
write on a full disk should leave the previous one intact rather than
half a document. Same directory because rename is only atomic within a
filesystem. Parent directories are created, so `-o out/nested/digest.md`
works without a prior mkdir.

Host resolution gains a third tier as a consequence of wiring it up: a host
the user named out loud (URL, `-R host/owner/repo`, `--hostname`) beats one
we inferred from a git remote, which in turn beats GH_HOST. The two-tier
version could not express the inferred case at all.

Exit codes are u8 rather than i32 now that ExitCode is what consumes them.

Verified against a live pull request: full and filtered output, file and
stdout targets, and the repo/not-found/usage exit codes.

+++ src/error.rs
@@ -35,7 +35,7 @@ pub enum Error {
 pub type Result<T> = std::result::Result<T, Error>;
 
 impl Error {
-    pub fn exit_code(&self) -> i32 {
+    pub fn exit_code(&self) -> u8 {
         match self {
             Error::Repo(_) => 3,
             Error::Auth(_) => 4,
